# docs/api/openapi.yaml
openapi: "3.0.3"
info:
  title: "Compute Platform API"
  version: "2024-01-15"
  description: >
    VM control plane API for instance lifecycle, images, volumes, snapshots,
    SSH keys, projects, and VPC networking.

    All mutating endpoints accept an optional `Idempotency-Key` header for
    retry-safe request deduplication.

    Authentication: `X-Principal-ID` header (Phase 1). Phase 2 targets
    SigV4 signed requests.

    All error responses follow the registered error envelope
    (see API_ERROR_CONTRACT_V1.md).

    ## Unsupported / Partially implemented

    - VPC/network endpoints: handler code is implemented and tested, but the
      routes are NOT yet wired into the production HTTP mux. Available only
      through the independent `NetworkHandlers` test seam.
    - Instance networking (NICs, SGs) at launch: gateway handler code exists
      but VPC network integration is incomplete.
    - Quota enforcement: count-based model active; reservation-based deferred.
    - RBAC: Phase 1 ownership model only (owner/principal). Phase 2 IAM not
      implemented.
    - KVM/real-machine runtime gate: all documented behaviour is proven at the
      handler seam via unit + integration tests. Hardware-gate validation
      requires Linux/KVM environment and is opt-in.
    - Image trust/validation pipeline: platform signature validation exists;
      user image import validation pipeline is minimal.
...
      (mTLS-protected) but not documented as a public endpoint — it is an
...
paths:
...
  /v1/images:
    get:
      summary: "List images"
      operationId: "listImages"
      responses:
        "200":
          description: "Image list (visible to principal)"
    post:
      summary: "Create image (from snapshot or import)"
      operationId: "createImage"
      responses:
        "202":
          description: "Image creation accepted (async)"
        "400":
          description: "Invalid request (missing source_type, etc.)"
  /v1/images/{id}:
    parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
    get:
      summary: "Describe image"
      operationId: "describeImage"
      responses:
        "200":
          description: "Image details"
        "404":
          description: "Not found or not accessible"
...
  /v1/images/{id}/grants:
    parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "string"
    get:
      summary: "List image share grants"
      operationId: "listImageGrants"
      responses:
        "200":
          description: "List of principals granted access"
    post:
      summary: "Grant image access to a principal"
      operationId: "grantImageAccess"
      responses:
        "200":
          description: "Grant created"
...